Forgot to add files in last commit

+++ downloader.py
@@ -2,19 +2,37 @@ import hathitrust_api.data_api as da
 import os
 import sys
 import math
+import json
+import time
+from requests.exceptions import HTTPError, RequestException
 
 
-def create_folders():
+def create_folders(base):
     """Creates folders for the texts.
     Each folder is in the format YYYY-XXXX, where XXXX is YYYY+25 years"""
     for i in range(1750, 1925, 25):
-        foldername = "Texts/{}-{}".format(i, i + 25)
+        foldername = base + "/{}-{}".format(i, i + 25)
         if not os.path.exists(foldername):
             os.makedirs(foldername)
 
     # Create N/A category
-    if not os.path.exists("Texts/NA"):
-        os.makedirs("Texts/NA")
+    if not os.path.exists(base + "/NA"):
+        os.makedirs(base + "/NA")
+
+
+def getnumpages(data_api, id):
+    """Downloads meta and return the number of pages in the book"""
+    while True:
+        try:
+            meta = data_api.getmeta(id, json=True)
+            break
+        except HTTPError:
+            print("Request for meta for book failed. Trying again.")
+            time.sleep(5)
+            pass
+
+    meta_parsed = json.loads(meta)
+    return int(meta_parsed["htd:numpages"])
 
 
 def download_book(data_api, id, book_date_id):
@@ -27,26 +45,61 @@ def download_book(data_api, id, book_date_id):
     # Find folder to save text
     folder = "Texts/"
     if date == "N/A":
-        folder += "NA"
+        folder += "NA/" + str(id)
     else:
         rounded = math.floor(int(date) / 25) * 25
         folder += str(rounded) + "-" + str(rounded+25) + "/" + str(id)
 
-    # Download book
-    ocrtext = data_api.getdocumentocr(hathi_id)
-
-    # Save book to folder
+    # Create folder if we have not already
     if not os.path.exists(folder):
         os.makedirs(folder)
 
-    for i in range(0, len(ocrtext)):
+    # Get number of pages in the book
+    numpages = getnumpages(data_api, hathi_id)
+
+    # Check if we already downloaded all pages of this book
+    if os.path.exists(folder + "/{}.txt".format(numpages)) or os.path.exists(folder + "/parsed.txt"):
+        print("Already downloaded book " + str(id) + " in folder " + folder)
+        return
+
+    # Download book, page by page
+    print("Now downloading book " + str(id) + " in folder " + folder)
+    print("---> Pages downloaded (Total={}): ".format(numpages))
+    for i in range(1, numpages + 1):
         filename = folder + "/{}.txt".format(i)
+
+        # If file already exists we don't re-download
+        if os.path.exists(filename):
+            continue
+
+        attempts = 0
+        while True:
+            try:
+                ocrpage = data_api.getpageocr(hathi_id, i)
+                break
+            except (HTTPError, RequestException) as e:
+                # We attempt to download the page a max of 10 times, then exit with error if we are still failing
+                # Note there are lots of issues with the HathiTrust rejecting requests after a few page downloads
+                attempts += 1
+                time.sleep(5)  # Wait a few seconds before trying again
+                if attempts <= 100:
+                    print("Failed to download page " + str(i) + " with " + str(attempts) + " tries. Trying again.")
+                    pass
+                else:
+                    raise HTTPError("Failed to download book " + str(id) + " at page " + str(i)) from e
         f = open(filename, 'wb')
-        f.write(ocrtext[i])
+        f.write(ocrpage)
         f.close()
 
+        print(str(i), end=' ', flush=True)
+    print()
+
+# ------------- Main script begins --------------- #
 # Create folders if they have not been already
-create_folders()
+create_folders("Texts")
+create_folders("Processed")
+create_folders("Combined")
+create_folders("Samples")
 
 # Create data_api instance to create connection to HathiTrust
 oauth_key = sys.argv[1]
@@ -54,6 +107,9 @@ oauth_secret_key = sys.argv[2]
 data_api = da.DataAPI(oauth_key, oauth_secret_key)
 
 # Download the books according to Bamman et al. (2017)
-ids = open("stratified.txt").read().split("\n")[1:2]
+ids = open("stratified.txt").read().split("\n")[1:]
 for i in range(0, len(ids)):
+    if i == 19 or i == 31:
+        continue
+
     download_book(data_api, i, ids[i])
